+++ src/shared/components/palette/case-link/write-case-link-field.component.spec.ts
@@ -2,11 +2,18 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { WriteCaseLinkFieldComponent } from './write-case-link-field.component';
 import { CUSTOM_ELEMENTS_SCHEMA, DebugElement } from '@angular/core';
 import { CaseField, FieldType } from '../../../domain/definition';
-import { ReactiveFormsModule } from '@angular/forms';
+import { FormGroup, ReactiveFormsModule } from '@angular/forms';
 import { PaletteUtilsModule } from '../utils';
 import { RouterTestingModule } from '@angular/router/testing';
 import { LinkedCasesService } from './services';
 import { CaseLink } from './domain';
+import { CaseEditPageComponent } from '../../case-editor/case-edit-page/case-edit-page.component';
+import { of } from 'rxjs';
+import { CaseEventData, Draft } from '../../../domain';
+import { FieldTypeSanitiser, FormValueService, FormErrorService, CaseFieldService, FormValidatorsService } from '../../../services';
+import { PageValidationService, Wizard, WizardPage } from '../../case-editor';
+import { PaletteService } from '../palette.service';
+import createSpyObj = jasmine.createSpyObj;
 
 const VALUE = {
   CaseReference: '1234-5678-1234-5678'
@@ -14,7 +21,7 @@ const VALUE = {
 const FIELD_ID = 'NewCaseLink';
 const FIELD_TYPE: FieldType = {
   id: 'CaseLink',
-  type: 'Complex',
+  type: 'Collection',
 };
 const CASE_REFERENCE: CaseField = <CaseField>({
   id: 'CaseReference',
@@ -60,12 +67,84 @@ const linkedCasesService = {
   linkedCases: linkedCases
 };
 
+class FieldTestComponent {}
+
+function createWizardPage(fields: CaseField[], isMultiColumn = false, order = 0): WizardPage {
+  const wp: WizardPage = new WizardPage();
+  wp.case_fields = fields;
+  wp.label = 'Test Label';
+  wp.getCol1Fields = () => fields;
+  wp.getCol2Fields = () => fields;
+  wp.isMultiColumn = () => isMultiColumn;
+  wp.order = order;
+  return wp;
+}
+
+function createCaseField(id: string, value: any, display_context = 'READONLY'): CaseField {
+  const cf = new CaseField();
+  cf.id = id;
+  cf.value = value;
+  cf.display_context = display_context;
+  return cf;
+}
+
 describe('WriteCaseLinkFieldComponent', () => {
+  const FORM_GROUP: FormGroup = new FormGroup({});
+  let caseEditPageComponent: CaseEditPageComponent;
+  let formValidatorService: any
   let component: WriteCaseLinkFieldComponent;
   let fixture: ComponentFixture<WriteCaseLinkFieldComponent>;
   let de: DebugElement;
+  const caseField2 = new CaseField();
+  let route: any;
+  const fieldTypeSanitiser = new FieldTypeSanitiser();
+  const formValueService = new FormValueService(fieldTypeSanitiser);
+  const formErrorService = new FormErrorService();
+  const caseFieldService = new CaseFieldService();
+  const pageValidationService = new PageValidationService(caseFieldService);
+  let dialog: any;
+  let paletteService: any;
+  let caseEditComponentStub: any;
+  const wizardPage = createWizardPage([createCaseField('field1', 'field1Value')], false, 0);
+  const WIZARD = new Wizard([wizardPage]);
+  const caseField1 = new CaseField();
+  const firstPage = new WizardPage();
+  let cancelled: any;
+  const someObservable = {
+    'subscribe': () => new Draft()
+  };
 
   beforeEach(async(() => {
+    formValidatorService = createSpyObj<FormValidatorsService>('formValidatorService', ['addValidators']);
+    paletteService = createSpyObj<PaletteService>('paletteService', [
+      'getFieldComponentClass'
+    ]);
+    paletteService.getFieldComponentClass.and.returnValue(FieldTestComponent);
+    caseEditComponentStub = {
+      'form': FORM_GROUP,
+      'wizard': WIZARD,
+      'data': '',
+      'eventTrigger': {'case_fields': [caseField1], 'name': 'Test event trigger name', 'can_save_draft': true},
+      'hasPrevious': () => true,
+      'getPage': () => firstPage,
+      'first': () => true,
+      'next': () => true,
+      'previous': () => true,
+      'cancel': () => undefined,
+      'cancelled': cancelled,
+      'validate': (caseEventData: CaseEventData) => of(caseEventData),
+      'saveDraft': (_: CaseEventData) => of(someObservable),
+      'caseDetails': {'case_id': '1234567812345678', 'tabs': [], 'metadataFields': [caseField2]},
+    };
+    route = {
+      params: of({id: 123}),
+      snapshot: {
+        queryParamMap: createSpyObj('queryParamMap', ['get'])
+      }
+    };
+    caseEditPageComponent = new CaseEditPageComponent(caseEditComponentStub,
+      route, formValueService, formErrorService, null, pageValidationService, dialog, caseFieldService);
+
     TestBed.configureTestingModule({
       imports: [
         ReactiveFormsModule,
@@ -77,13 +156,17 @@ describe('WriteCaseLinkFieldComponent', () => {
         WriteCaseLinkFieldComponent,
       ],
       providers: [
-        { provide: LinkedCasesService, useValue: linkedCasesService }]
+        { provide: LinkedCasesService, useValue: linkedCasesService },
+        { provide: CaseEditPageComponent, useValue: caseEditPageComponent },
+      ]
     })
     .compileComponents();
 
     fixture = TestBed.createComponent(WriteCaseLinkFieldComponent);
     component = fixture.componentInstance;
+    component.caseEditPageComponent = caseEditComponentStub;
     component.caseField = CASE_FIELD;
+    component.formGroup = FORM_GROUP;
     de = fixture.debugElement;
     fixture.detectChanges();
   }));
@@ -99,10 +182,4 @@ describe('WriteCaseLinkFieldComponent', () => {
     expect(component.validCaseReference('123456781234567890')).toBeFalsy();
     expect(component.validCaseReference('1234Invalid')).toBeFalsy();
   });
-
-  it('should set retain_hidden_value to true for all sub-fields that are part of a CaseLink field', () => {
-    expect(component.caseField.field_type.complex_fields.length).toEqual(1);
-    expect(component.caseField.field_type.complex_fields[0].retain_hidden_value).toEqual(true);
-  });
-
 });