+++ src/shared/components/palette/case-link/write-case-link-field.component.ts
@@ -1,45 +1,49 @@
 import { Component, Input, OnInit, ViewChild } from '@angular/core';
-import { AbstractControl, FormControl, FormGroup, ValidatorFn, Validators } from '@angular/forms';
-import { Router } from '@angular/router';
+import { AbstractControl, FormArray, FormControl, FormGroup, Validators } from '@angular/forms';
+import { CaseField } from '../../../domain/definition/case-field.model';
 import { CaseEditPageComponent } from '../../case-editor/case-edit-page/case-edit-page.component';
 import { AbstractFieldWriteComponent } from '../base-field/abstract-field-write.component';
 import { WriteComplexFieldComponent } from '../complex/write-complex-field.component';
+import { LinkedCasesEventTriggers } from './enums';
 import { LinkedCasesService } from './services';
 
 @Component({
   selector: 'ccd-write-case-link-field',
   templateUrl: 'write-case-link-field.html'
 })
+
 export class WriteCaseLinkFieldComponent extends AbstractFieldWriteComponent implements OnInit {
+
+  @Input()
+  public caseFields: CaseField[] = [];
+
+  @Input()
+  public formGroup: FormGroup;
+
+  public formArray: FormArray;
 
   @Input()
   public caseEditPageComponent: CaseEditPageComponent;
 
-  caseReferenceControl: AbstractControl;
-  caseLinkGroup: FormGroup;
-  containsCaseLinkCollection: boolean;
+  public caseReferenceControl: AbstractControl;
+  public caseLinkGroup: FormGroup;
+  public containsCaseLinkCollection: boolean;
 
   @ViewChild('writeComplexFieldComponent')
   writeComplexFieldComponent: WriteComplexFieldComponent;
 
-  constructor(private router: Router,
-    private readonly linkedCasesService: LinkedCasesService) {
+  constructor(private readonly linkedCasesService: LinkedCasesService) {
     super();
   }
 
   public ngOnInit(): void {
-    if (this.caseField.value) {
+    this.formArray = this.registerControl(new FormArray([]), true) as FormArray;
+    this.formArray['caseField'] = this.caseField;
+    if (!this.hasCaseLinkCollection()) {
       this.caseLinkGroup = this.registerControl(new FormGroup({
         'CaseReference': new FormControl(this.caseField.value.CaseReference, Validators.required),
       }), true) as FormGroup;
-    } else {
-      this.caseLinkGroup = this.registerControl(new FormGroup({
-        'CaseReference': new FormControl(null, Validators.required),
-      }), true) as FormGroup;
     }
-    this.caseReferenceControl = this.caseLinkGroup.controls['CaseReference'];
-    this.caseReferenceControl.setValidators(this.caseReferenceValidator());
-
     // Ensure that all sub-fields inherit the same value for retain_hidden_value as this parent; although a CaseLink
     // field uses the Complex type, it is meant to be treated as one field
     if (this.caseField && this.caseField.field_type.type === 'Complex') {
@@ -50,24 +54,15 @@ export class WriteCaseLinkFieldComponent extends AbstractFieldWriteComponent imp
     this.containsCaseLinkCollection = this.hasCaseLinkCollection();
   }
 
-  public submitLinkedCasses(): void {
-    this.caseReferenceControl.setValue(this.linkedCasesService.linkedCases.map((caseInfo) => caseInfo.caseReference)[0]);
-  }
-
-  private caseReferenceValidator(): ValidatorFn {
-    return (control: AbstractControl): { [key: string]: any } => {
-      if (control.value) {
-        if (this.validCaseReference(control.value)) {
-          return null;
-        }
-        return { 'error': 'Please use a valid 16 Digit Case Reference' };
-      } else {
-        if (control.touched) {
-          return { 'error': 'Please use a valid 16 Digit Case Reference' };
-        }
-      }
-      return null;
-    };
+  public submitLinkedCases(): void {
+    const formGroup = this.formGroup;
+    if (!this.linkedCasesService.isLinkedCasesEventTrigger) {
+      const unlinkedCaseRefereneIds = this.linkedCasesService.linkedCases.filter(item => item.unlink).map(item => item.caseReference);
+      this.formGroup.value.caseLinks = this.linkedCasesService.caseFieldValue
+                                        .filter(item => unlinkedCaseRefereneIds.indexOf(item.id) === -1);
+    } else if (formGroup.value && formGroup.value.caseLinks && this.linkedCasesService.linkedCases) {
+      this.formGroup.value.caseLinks = this.linkedCasesService.caseFieldValue;
+    }
   }
 
   public validCaseReference(valueString: string): boolean {
@@ -78,7 +73,10 @@ export class WriteCaseLinkFieldComponent extends AbstractFieldWriteComponent imp
   }
 
   public hasCaseLinkCollection(): boolean {
-    return this.caseField.field_type && this.caseField.field_type.type === 'Collection' &&
-            this.caseField.field_type.collection_field_type.id === 'CaseLink'
+    return (
+      this.caseField.field_type &&
+      this.caseField.field_type.type === 'Collection' &&
+      this.caseField.field_type.collection_field_type.id === 'CaseLink'
+    );
   }
 }